+++ src/release_notes.md
@@ -1,7 +1,8 @@
 # Release v1.4.5.0-NET5-RC5
 
 ## CLI
-- Added complex (multi-file) commands support (e.g. `css -selftest:run`)
+- Added complex (multi-file) commands support (e.g. `css -self-test-run`)
 - Build changes to support Linux distro
-- `-selftest` switch (wip)
 - Added `-vscode` help info
+- Added `css -server:restart` CLI parameter
+- Added detection of CLI host and build server running with different root privileges on Linux